requirements.txt
numpy>=1.21
pandas
gdown
scikit-learn
boto3
pytorch_lightning==1.6.5
pytorch-crf
tweet-preprocessor
seqeval
imbalanced-learn
textaugment
sentencepiece
texttable
latextable
colorama
termcolor
flair==0.6.1
gensim>=3.8.3
pytorch_transformers
faiss-gpu>=1.5.3
pyarrow==6.0.1
datasets>=1.18.3
docopt>=0.6.2
facenet-pytorch>=2.5.2
SPARQLWrapper>=1.8.5
tabulate>=0.8.9
tqdm>=4.49.0
seaborn>=0.11
spacy>=2.2.4
elasticsearch==7.7.1
transformers==4.30.2
optuna>=2.9.1
scikit-image
opencv-python>=4.5.3
humanize>=3.12
ra
easydict
LTNtorch
igraph
cairocffi
docopt
ranx
pytorch-metric-learning
jsonargparse
psutil
pycocotools
vector-quantize-pytorch
peft==0.2.0






